FMI Import/Core .jl Library Functions/Types

FMICore.FMU — Type

The mutable struct representing an abstract (version unknown) FMU.

FMICore.FMUSolution — Type

ToDo.

FMICore.FMUEvent — Type

ToDo.

FMICore.FMUExecutionConfiguration — Type

ToDo.

FMICore.FMULogLevel — Type

Log levels for non-standard infos, warnings and errors.

FMICore.logInfo — Function

Logs a message with level info if the log level allows it.

FMICore.logWarning — Function

Logs a message with level warn if the log level allows it.

FMICore.logError — Function

Logs a message with level error if the log level allows it.

FMIImport.loadBinary — Function

TODO