FMI.jl Library Functions

Many of the functions in this library are based on already defined functions of the FMIImport.jl library.

Simulate FMUs

FMI.fmiLoad — Function

fmiLoad(pathToFMU::String; unpackPath=nothing, type=nothing)

Load FMUs independent of the FMI version, currently supporting version 2.0.X and 3.0.

Arguments

pathToFMU::String: String that contains the paths of ziped and unziped FMU folders.

Keywords

unpackPath=nothing: Via optional argument unpackPath, a path to unpack the FMU can be specified (default: system temporary directory).

type::Union{CS, ME, SE} = nothing: Via type, a FMU type can be selected. If none of the unified type set is used, the default value type = nothing will be used.

Returns

Returns the instance of the FMU struct.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2Load, fmi3Load.

FMI.fmiSimulate — Function

fmiSimulate(str::Union{fmi2Struct, fmi3Struct}, tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing;
                tolerance::Union{Real, Nothing} = nothing,
                dt::Union{Real, Nothing} = nothing,
                solver = nothing,
                customFx = nothing,
                recordValues::fmi2ValueReferenceFormat = nothing,
                saveat = nothing,
                x0::Union{AbstractArray{<:Real}, Nothing} = nothing,
                setup::Union{Bool, Nothing} = nothing,
                reset::Union{Bool, Nothing} = nothing, # nothing = auto
                instantiate::Union{Bool, Nothing} = nothing,
                freeInstance::Union{Bool, Nothing} = nothing,
                terminate::Union{Bool, Nothing} = nothing,
                inputValueReferences::fmi2ValueReferenceFormat = nothing,
                inputFunction = nothing,
                parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing,
                dtmax::Union{Real, Nothing} = nothing,
                kwargs...)

Starts a simulation of the FMU instance for the matching FMU type, if both types are available, CS is preferred.

Arguments

str::Union{fmi2StructMD, fmi3StructMD}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2Struct = Union{FMU2, FMU2Component} More detailed: fmi3Struct = Union{FMU3, FMU3Instance}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing: Sets the time span as a tuple or the default value from the model description is used.

Keywords

tolerance::Union{Real, Nothing} = nothing: Real number to set the tolerance for any OED-solver

dt::Union{Real, Nothing} = nothing: Real number to set the step size of the OED-solver. Defaults to an automatic choice if the method is adaptive. More Info: DifferentialEquations.jl Documentation

solver = nothing: Any Julia-supported OED-solver (default is Tsit5). More Info: DifferentialEquations.jl Documentation

customFx = nothing: [deperecated] Ability to give a custom state derivative function ẋ=f(x,t)

recordValues::fmi2ValueReferenceFormat = nothing: AbstractArray of variables (strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues

saveat = []: Time points to save values at (interpolated). More Info: DifferentialEquations.jl Documentation

saveat = nothing: Time points to save values at (interpolated). More Info: DifferentialEquations.jl Documentation

x0::Union{AbstractArray{<:Real}, Nothing} = nothing: Stores the specific value of fmi2ScalarVariable containing the modelVariables with the identical fmi2ValueReference to the input variable vr (vr = vrs[i]). And is therefore passed within prepareSolveFMU to fmi2Set , to set the start state.

setup::Bool = true: Boolean value if FMU is to be set up (default: setup=true).

reset::Union{Bool, Nothing} = nothing: Boolean value that determines whether the FMU should be reset before simulation (default: reset:=auto).

instantiate::Union{Bool, Nothing} = nothing: Boolean value that decides whether to create a new instance of the specified fmu.

freeInstance::Union{Bool, Nothing} = nothing: Boolean value that determines whether to dispose of the given instance, unload the loaded model, and free the allocated memory and other resources allocated by the FMU interface functions.

terminate::Union{Bool, Nothing} = nothing: Boolean value that tells the FMU that the simulation run will be aborted.

inputValueReferences::fmi2ValueReferenceFormat = nothing: AbstractArray of input variables (strings or variableIdentifiers) to set at every simulation step

inputFunction = nothing: Function to retrieve the values to set the inputs to

parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing: Dictionary of parameter variables (strings or variableIdentifiers) and values (Real, Integer, Boolean, String) to set parameters during initialization

dtmax::Union{Real, Nothing} = nothing: Real number for setting maximum dt for adaptive timestepping for the ODE solver. The default values are package dependent. More Info: DifferentialEquations.jl Documentation

callbacks = []: custom callbacks to add.

showProgress::Bool = true: Boolean value that determines whether a progress bar is generated for a task

kwargs...: Further parameters of already defined functions solve(args..., kwargs...) from the library DifferentialEquations.jl

Returns

success::Bool for CS-FMUs

ODESolution for ME-FMUs

if keyword recordValues is set, a tuple of type (success::Bool, DiffEqCallbacks.SavedValues) for CS-FMUs

if keyword recordValues is set, a tuple of type (ODESolution, DiffEqCallbacks.SavedValues) for ME-FMUs

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2Simulate, fmi2SimulateME, fmi2SimulateCS, fmi2Struct, FMU2, FMU2Component.

FMI.fmiSimulateCS — Function

fmiSimulateCS(str::Union{fmi2Struct,fmi3Struct}, tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing;
                    tolerance::Union{Real, Nothing} = nothing,
                    dt::Union{Real, Nothing} = nothing,
                    solver = nothing,
                    customFx = nothing,
                    recordValues::fmi2ValueReferenceFormat = nothing,
                    saveat = [],
                    setup::Bool = true,
                    reset::Union{Bool, Nothing} = nothing, # nothing = auto
                    instantiate::Union{Bool, Nothing} = nothing,
                    freeInstance::Union{Bool, Nothing} = nothing,
                    terminate::Union{Bool, Nothing} = nothing,
                    inputValueReferences::fmi2ValueReferenceFormat = nothing,
                    inputFunction = nothing,
                    showProgress::Bool=true,
                    parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing,
                    dtmax::Union{Real, Nothing} = nothing,
                    kwargs...)

Starts a simulation of the Co-Simulation FMU instance.

Arguments

str::Union{fmi2StructMD, fmi3StructMD}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2Struct = Union{FMU2, FMU2Component} More detailed: fmi3Struct = Union{FMU3, FMU3Instance}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing: Sets the time span as a tuple or the default value from the model description is used.

Keywords

tolerance::Union{Real, Nothing} = nothing: Real number to set the tolerance for any OED-solver

dt::Union{Real, Nothing} = nothing: Real number to set the step size of the OED-solver. Defaults to an automatic choice if the method is adaptive. More Info: DifferentialEquations.jl Documentation

solver = nothing: Any Julia-supported OED-solver (default is Tsit5). More Info: DifferentialEquations.jl Documentation

customFx = nothing: [deperecated] Ability to give a custom state derivative function ẋ=f(x,t)

recordValues::fmi2ValueReferenceFormat = nothing: AbstractArray of variables (strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues

saveat = []: Time points to save values at (interpolated). More Info: DifferentialEquations.jl Documentation

setup::Bool = true: Boolean, if FMU should be setup (default: setup=true)

reset::Union{Bool, Nothing} = nothing: Boolean, if FMU should be reset before simulation (default: reset:=auto)

instantiate::Union{Bool, Nothing} = nothing: Boolean value that decides whether to create a new instance of the specified fmu.

freeInstance::Union{Bool, Nothing} = nothing: Boolean value that determines whether to dispose of the given instance, unload the loaded model, and free the allocated memory and other resources allocated by the FMU interface functions.

terminate::Union{Bool, Nothing} = nothing: Boolean value that tells the FMU that the simulation run will be aborted.

inputValueReferences::fmi2ValueReferenceFormat = nothing: AbstractArray of input variables (strings or variableIdentifiers) to set at every simulation step

inputFunction = nothing: Function to retrieve the values to set the inputs to

showProgress::Bool = true: Boolean value that determines whether a progress bar is generated for a task

parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing: Dictionary of parameter variables (strings or variableIdentifiers) and values (Real, Integer, Boolean, String) to set parameters during initialization

dtmax::Union{Real, Nothing} = nothing: Real number for setting maximum dt for adaptive timestepping for the ODE solver. The default values are package dependent. More Info: DifferentialEquations.jl Documentation

kwargs...: Further parameters of already defined functions solve(args..., kwargs...) from the library DifferentialEquations.jl

Returns

If keyword recordValues is not set, a boolean success is returned (simulation success).

If keyword recordValues is set, a tuple of type (true, DiffEqCallbacks.SavedValues) or (false, nothing).

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2SimulateCS, fmi2Simulate, fmi2SimulateME, fmi2Struct, FMU2, FMU2Component, fmi3SimulateCS, fmi3Simulate, fmi3SimulateME, fmi3Struct, FMU3, FMU3Instance.

FMI.fmiSimulateME — Function

fmiSimulateME(str::Union{fmi2Struct,fmi3Struct}, tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing;
                tolerance::Union{Real, Nothing} = nothing,
                dt::Union{Real, Nothing} = nothing,
                solver = nothing,
                customFx = nothing,
                recordValues::fmi2ValueReferenceFormat = nothing,
                saveat = nothing,
                x0::Union{AbstractArray{<:Real}, Nothing} = nothing,
                setup::Union{Bool, Nothing} = nothing,
                reset::Union{Bool, Nothing} = nothing,
                instantiate::Union{Bool, Nothing} = nothing,
                freeInstance::Union{Bool, Nothing} = nothing,
                terminate::Union{Bool, Nothing} = nothing,
                inputValueReferences::fmi2ValueReferenceFormat = nothing,
                inputFunction = nothing,
                parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing,
                dtmax::Union{Real, Nothing} = nothing,
                callbacks = [],
                showProgress::Bool = true,
                kwargs...)

Simulates a FMU instance for the given simulation time interval.

Arguments

str::Union{fmi2StructMD, fmi3StructMD}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2Struct = Union{FMU2, FMU2Component} More detailed: fmi3Struct = Union{FMU3, FMU3Instance}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing: Sets the time span as a tuple or the default value from the model description is used.

Keywords

tolerance::Union{Real, Nothing} = nothing: Real number to set the tolerance for any OED-solver

dt::Union{Real, Nothing} = nothing: Real number to set the step size of the OED-solver. Defaults to an automatic choice if the method is adaptive. More Info: DifferentialEquations.jl Documentation

solver = nothing: Any Julia-supported OED-solver (default is Tsit5). More Info: DifferentialEquations.jl Documentation

customFx = nothing: [deperecated] Ability to give a custom state derivative function ẋ=f(x,t)

recordValues::fmi2ValueReferenceFormat = nothing: AbstractArray of variables (strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues

saveat = nothing: Time points to save values at (interpolated). More Info: DifferentialEquations.jl Documentation

x0::Union{AbstractArray{<:Real}, Nothing} = nothing: Stores the specific value of fmi2ScalarVariable containing the modelVariables with the identical fmi2ValueReference to the input variable vr (vr = vrs[i]). And is therefore passed within prepareSolveFMU to fmi2Set , to set the start state.

setup::Bool = true: Boolean, if FMU should be setup (default: setup=true)

reset::Union{Bool, Nothing} = nothing: Boolean, if FMU should be reset before simulation (default: reset:=auto)

instantiate::Union{Bool, Nothing} = nothing: Boolean value that decides whether to create a new instance of the specified fmu.

freeInstance::Union{Bool, Nothing} = nothing: Boolean value that determines whether to dispose of the given instance, unload the loaded model, and free the allocated memory and other resources allocated by the FMU interface functions.

terminate::Union{Bool, Nothing} = nothing: Boolean value that tells the FMU that the simulation run will be aborted.

inputValueReferences::fmi2ValueReferenceFormat = nothing: AbstractArray of input variables (strings or variableIdentifiers) to set at every simulation step

inputFunction = nothing: Function to retrieve the values to set the inputs to

parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing: Dictionary of parameter variables (strings or variableIdentifiers) and values (Real, Integer, Boolean, String) to set parameters during initialization

dtmax::Union{Real, Nothing} = nothing: Real number for setting maximum dt for adaptive timestepping for the ODE solver. The default values are package dependent. More Info: DifferentialEquations.jl Documentation

callbacks = []: custom callbacks to add.

showProgress::Bool = true: Boolean value that determines whether a progress bar is generated for a task

kwargs...: Further parameters of already defined functions solve(args..., kwargs...) from the library DifferentialEquations.jl

Returns

If keyword recordValues is not set, a struct of type ODESolution.

If keyword recordValues is set, a tuple of type (ODESolution, DiffEqCallbacks.SavedValues).

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2SimulateME fmi2SimulateCS, fmi2Simulate, fmi2Struct, FMU2, FMU2Component.

FMI.fmiUnload — Function

fmiUnload(fmu::Union{FMU2, FMU3})

Unloads the FMU and all its instances and frees the allocated memory.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

fmu::FMU3: Mutable struct representing a FMU and all it instantiated instances in the FMI 3.0 Standard.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2Unload, fmi3Unload.

FMI.fmiReload — Function

fmiReload(fmu::Union{FMU2, FMU3})

Reloads the FMU-binary. This is useful, if the FMU does not support a clean reset implementation.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

fmu::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2Reload, fmi3Reload.

Handling Value References

FMI.fmiStringToValueReference — Function

fmiStringToValueReference(dataStruct::Union{FMU2, fmi2ModelDescription, FMU3, fmmi3ModelDescription}, identifier::Union{String, AbstractArray{String}})

Returns the ValueReference coresponding to the variable identifier.

Arguments

dataStruct::Union{FMU2, fmi2ModelDescription, FMU3, fmmi3ModelDescription}: Model of the type FMU2/FMU3 or the Model Description of fmi2/fmi3. Same for Model of type FMU3 or the Model Description of fmi3

identifier::Union{String, AbstractArray{String}}: Variable identifier in type String or as a 1-dimensional AbstractArray containing elements of type String

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2StringToValueReference, fmi3StringToValueReference.

FMI.fmiGetStartValue — Function

fmiGetStartValue(s::fmi2Struct, vr::fmi2ValueReferenceFormat)

Returns the start/default value for a given value reference.

Arguments

md::fmi2ModelDescription: Struct which provides the static information of ModelVariables.

vrs::fmi2ValueReferenceFormat = md.valueReferences: wildcards for how a user can pass a fmi[X]ValueReference (default = md.valueReferences)

More detailed: fmi2ValueReferenceFormat = Union{Nothing, String, Array{String,1}, fmi2ValueReference, Array{fmi2ValueReference,1}, Int64, Array{Int64,1}, Symbol}

Returns

starts::Array{fmi2ValueReferenceFormat}: start/default value for a given value reference

External/additional functions

FMI.fmiSet — Function

fmiSet(fmu::FMU3, args...; kwargs...)

Wrapper for fmiSet() in FMIImport/FMI3_ext.jl

fmiSet(str::fmi2Struct, comp::FMU2Component, vrs::fmi2ValueReferenceFormat, srcArray::AbstractArray; filter=nothing)

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

srcArray::AbstractArray: Stores the specific value of fmi2ScalarVariable containing the modelVariables with the identical fmi2ValueReference to the input variable vr (vr = vrs[i]). srcArray has the same length as vrs.

Keywords

filter=nothing: whether the individual values of "fmi2ScalarVariable" are to be stored

Returns

retcodes::Array{fmi2Status}: Returns an array of length length(vrs) with Type fmi2Status. Type fmi2Status is an enumeration and indicates the success of the function call.

More detailed:

fmi2OK: all well

fmi2Warning: things are not quite right, but the computation can continue

fmi2Discard: if the slave computed successfully only a subinterval of the communication step

fmi2Error: the communication step could not be carried out at all

fmi2Fatal: if an error occurred which corrupted the FMU irreparably

fmi2Pending: this status is returned if the slave executes the function asynchronously

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.23]: 2.1.6 Initialization, Termination, and Resetting an FMU

FMISpec2.0.2[p.18]: 2.1.3 Status Returned by Functions

FMI.fmiGet — Function

fmiGet(fmu::FMU3, args...; kwargs...)

Wrapper for fmiGet() in FMIImport/FMI3_ext.jl

fmiGet(str::fmi2Struct, comp::FMU2Component, vrs::fmi2ValueReferenceFormat)

Returns the specific value of fmi2ScalarVariable containing the modelVariables with the identical fmi2ValueReference in an array.

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

vrs::fmi2ValueReferenceFormat: wildcards for how a user can pass a fmi[X]ValueReference

More detailed: fmi2ValueReferenceFormat = Union{Nothing, String, Array{String,1}, fmi2ValueReference, Array{fmi2ValueReference,1}, Int64, Array{Int64,1}, Symbol}

Returns

dstArray::Array{Any,1}(undef, length(vrs)): Stores the specific value of fmi2ScalarVariable containing the modelVariables with the identical fmi2ValueReference to the input variable vr (vr = vrs[i]). dstArray is a 1-Dimensional Array that has the same length as vrs.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.23]: 2.1.6 Initialization, Termination, and Resetting an FMU

FMISpec2.0.2[p.18]: 2.1.3 Status Returned by Functions

FMI.fmiGet! — Function

fmiGet!(fmu::FMU3, args...; kwargs...)

Wrapper for fmiGet!() in FMIImport/FMI3_ext.jl

fmiGet!(str::fmi2Struct, comp::FMU2Component, vrs::fmi2ValueReferenceFormat, dstArray::AbstractArray)

Stors the specific value of fmi2ScalarVariable containing the modelVariables with the identical fmi2ValueReference in an array.

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

dstArray::AbstractArray: Stores the specific value of fmi2ScalarVariable containing the modelVariables with the identical fmi2ValueReference to the input variable vr (vr = vrs[i]). dstArray has the same length as vrs.

Returns

retcodes::Array{fmi2Status}: Returns an array of length length(vrs) with Type fmi2Status. Type fmi2Status is an enumeration and indicates the success of the function call.

More detailed:

fmi2OK: all well

fmi2Warning: things are not quite right, but the computation can continue

fmi2Discard: if the slave computed successfully only a subinterval of the communication step

fmi2Error: the communication step could not be carried out at all

fmi2Fatal: if an error occurred which corrupted the FMU irreparably

fmi2Pending: this status is returned if the slave executes the function asynchronously

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.23]: 2.1.6 Initialization, Termination, and Resetting an FMU

FMISpec2.0.2[p.18]: 2.1.3 Status Returned by Functions

FMI.fmiGetNumberOfStates — Function

fmiGetNumberOfStates(str::Union{fmi2Struct, fmi3Struct})

Returns the number of states of the FMU.

Arguments

str::Union{fmi2Struct, fmi3Struct}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2Struct = Union{FMU2, FMU2Component} More detailed: fmi3StructMD = Union{FMU3, FMU3Instance, fmi3ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

Returns

Returns the length of the md.valueReferences::Array{fmi2ValueReference} corresponding to the number of states of the FMU.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2GetNumberOfStates, fmi2Struct, FMU2, FMU2Component, , fmi3Struct, FMU3, FMU3Instance.

FMI.fmiCanGetSetState — Function

fmiCanGetSetState(str::Union{fmi2StructMD, fmi3StructMD})

Returns true, if the FMU supports the getting/setting of states

Arguments

str::Union{fmi2StructMD, fmi3StructMD}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription} More detailed: fmi3StructMD = Union{FMU3, FMU3Instance, fmi3ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct witch provides the static information of ModelVariables.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

str::fmi3ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

::Bool: The function fmi2CanGetSetState returns True, if the FMU supports the getting/setting of states.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2CanGetSetState, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription, fmi3CanGetSetState, fmi3StructMD, FMU3, FMU3Instance, fmi3ModelDescription.

FMI.fmiGetState — Function

fmiGetState(str::fmi2Struct)

Makes a copy of the internal FMU state and returns a pointer to this copy.

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

Returns

Return state is a pointer to a copy of the internal FMU state.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions (fmi2TypesPlatform.h)

FMISpec2.0.2[p.25]: 2.1.8 Getting and Setting the Complete FMU State

See also fmi2GetFMUstate, fmi2Struct, FMU2, FMU2Component.

FMI.fmiSetState — Function

fmiSetState(str::fmi2Struct, c::FMU2Component, FMUstate::fmi2FMUstate)

Sets the FMU to the given state

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

FMUstate::fmi2FMUstate: Argument FMUstate is a pointer to a data structure in the FMU that saves the internal FMU state of the actual or a previous time instant.

Returns

status::fmi2Status: Return status is an enumeration of type fmi2Status and indicates the success of the function call.

More detailed:

fmi2OK: all well

fmi2Warning: things are not quite right, but the computation can continue

fmi2Discard: if the slave computed successfully only a subinterval of the communication step

fmi2Error: the communication step could not be carried out at all

fmi2Fatal: if an error occurred which corrupted the FMU irreparably

fmi2Pending: this status is returned if the slave executes the function asynchronously

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions (fmi2TypesPlatform.h)

FMISpec2.0.2[p.16]: 2.1.3 Status Returned by Functions

FMISpec2.0.2[p.25]: 2.1.8 Getting and Setting the Complete FMU State

See also fmi2GetFMUstate, fmi2Struct, FMU2, FMU2Component.

FMI.fmiFreeState! — Function

fmiFreeState!(str::fmi2Struct, c::FMU2Component, state::fmi2FMUstate)
fmiFreeState!(str::fmi2Struct, c::FMU2Component, FMUstate::Ref{fmi2FMUstate})

Free the memory for the allocated FMU state

Arguments

str::fmi2Struct: Representative for an FMU in the FMI 2.0.2 Standard.

More detailed: fmi2Struct = Union{FMU2, FMU2Component}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

state::fmi2FMUstate: Argument state is a pointer to a data structure in the FMU that saves the internal FMU state of the actual or a previous time instant.

FMUstate::Ref{fmi2FMUstate}: Argument FMUstate is an object that safely references data of type fmi3FMUstate wich is a pointer to a data structure in the FMU that saves the internal FMU state of the actual or a previous time instant.

Returns

Return singleton instance of type Nothing, if there is no value to return (as in a C void function) or when a variable or field holds no value.

status::fmi2Status: Return status is an enumeration of type fmi2Status and indicates the success of the function call.

More detailed:

fmi2OK: all well

fmi2Warning: things are not quite right, but the computation can continue

fmi2Discard: if the slave computed successfully only a subinterval of the communication step

fmi2Error: the communication step could not be carried out at all

fmi2Fatal: if an error occurred which corrupted the FMU irreparably

fmi2Pending: this status is returned if the slave executes the function asynchronously

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions (fmi2TypesPlatform.h)

FMISpec2.0.2[p.16]: 2.1.3 Status Returned by Functions

FMISpec2.0.2[p.25]: 2.1.8 Getting and Setting the Complete FMU State

See also fmi2FreeFMUstate!,fmi2FMUstate, fmi2Struct, FMU2, FMU2Component.

FMI.fmiGetDependencies — Function

fmiGetDependencies(fmu::FMU2)

Building dependency matrix dim x dim for fast look-ups on variable dependencies (dim is number of states).

Arguments

fmu::FMU2: Mutable Struct representing a FMU.

Retruns

fmu.dependencies::Matrix{Union{fmi2DependencyKind, Nothing}}: Returns the FMU's dependency-matrix for fast look-ups on dependencies between value references. Entries are from type fmi2DependencyKind.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2GetDependencies.

FMI.fmiProvidesDirectionalDerivative — Function

fmiProvidesDirectionalDerivative(str::Union{fmi2StructMD, fmi3StructMD})

Returns true, if the FMU provides directional derivatives

Arguments

str::Union{fmi2StructMD, fmi3StructMD}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription} More detailed: fmi3StructMD = Union{FMU3, FMU3Instance, fmi3ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct witch provides the static information of ModelVariables.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

str::fmi3ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

::Bool: The function fmi2ProvidesDirectionalDerivative returns True, if the FMU provides directional derivatives.

See also fmi2ProvidesDirectionalDerivative, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription, fmi3ProvidesDirectionalDerivatives, fmi3StructMD, FMU3, FMU3Instance, fmi3ModelDescription.

FMI.fmiInfo — Function

fmiInfo(str::fmi2Struct)

Prints FMU-specific information into the REPL.

Arguments

str::Union{fmi2Struct, fmi3Struct}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2Struct = Union{FMU2, FMU2Component} More detailed: fmi3StructMD = Union{FMU3, FMU3Instance, fmi3ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

Returns

Prints FMU related information.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec3.0 Link: https://fmi-standard.org/

See also fmi2Info, fmi2Struct, FMU2, FMU2Component, fmi3Info, fmi3Struct, FMU3, FMU3Instance.

FMI.fmiGetModelName — Function

fmiGetModelName(str::Union{fmi2StructMD, fmi3StructMD})

Returns the tag 'modelName' from the model description.

Arguments

str::Union{fmi2StructMD, fmi3StructMD}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

Union{fmi2StructMD, fmi3StructMD} = Union{FMU2, FMU2Component, fmi2ModelDescription, FMU3, FMU3Instance, fmi3ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct wich provides the static information of ModelVariables.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

str::fmi3ModelDescription: Struct witch provides the static information of ModelVariables.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also FMU2, FMU2Component, fmi2ModelDescription, FMU3, FMU3Instance, fmi3ModelDescription.

FMI.fmiGetGUID — Function

fmiGetGUID(str::fmi2StructMD)

Returns the tag 'guid' from the model description.

Arguments

str::fmi2StructMD: Representative for an FMU in the FMI 2.0.2 Standard. More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct wich provides the static information of ModelVariables.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2GetGUID, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription.

FMI.fmiIsCoSimulation — Function

fmiIsCoSimulation(str::Union{fmi2StructMD, fmi3StructMD})

Returns true, if the FMU supports co simulation

Arguments

str::Union{fmi2StructMD, fmi3StructMD}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription} More detailed: fmi3StructMD = Union{FMU3, FMU3Instance, fmi3ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct witch provides the static information of ModelVariables.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

str::fmi3ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

::Bool: The function fmi2IsCoSimulation returns True, if the FMU supports co simulation

See also fmi2IsCoSimulation, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription, fmi3IsCoSimulation, fmi3StructMD, FMU3, FMU3Instance, fmi3ModelDescription.

FMI.fmiIsModelExchange — Function

fmiIsModelExchange(str::Union{fmi2StructMD, fmi3StructMD})

Returns true, if the FMU supports model exchange

Arguments

str::Union{fmi2StructMD, fmi3StructMD}: Representative for an FMU in the FMI 2.0.2 Standard or FMI 3.0 Standard. Other notation:

More detailed: fmi2StructMD = Union{FMU2, FMU2Component, fmi2ModelDescription} More detailed: fmi3StructMD = Union{FMU3, FMU3Instance, fmi3ModelDescription}

str::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

str::FMU2Component: Mutable struct represents an instantiated instance of an FMU in the FMI 2.0.2 Standard.

str::fmi2ModelDescription: Struct witch provides the static information of ModelVariables.

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

str::fmi3ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

::Bool: The function fmi2IsModelExchange returns True, if the FMU supports model exchange.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

See also fmi2IsModelExchange, fmi2StructMD, FMU2, FMU2Component, fmi2ModelDescription, fmi3IsModelExchange, fmi3StructMD, FMU3, FMU3Instance, fmi3ModelDescription.

FMI.fmiIsScheduledExecution — Function

fmiIsScheduledExecution(str::fmi3StructMD)

Returns true, if the FMU supports scheduled execution

Arguments

str::fmi3StructMD: Representative for an FMU in the FMI 3.0 Standard. Other notation:

More detailed: fmi3StructMD = Union{FMU3, FMU3Instance, fmi3ModelDescription}

str::FMU3: Mutable struct representing an FMU in the FMI 3.0 Standard.

str::FMU3Instance: Mutable struct represents a pointer to an FMU specific data structure that contains the information needed. Also in FMI 3.0 Standard.

str::fmi3ModelDescription: Struct witch provides the static information of ModelVariables.

Returns

::Bool: The function fmi3IsScheduledExecution returns True, if the FMU supports scheduled execution.

See also fmi3IsScheduledExecution, fmi3StructMD, FMU3, FMU3Instance, fmi3ModelDescription.

Visualize simulation results

FMI.fmiPlot — Function

fmiPlot(solution::FMUSolution; states::Union{Bool, Nothing}=nothing,
values::Union{Bool, Nothing}=nothing,
stateEvents::Union{Bool, Nothing}=nothing,
timeEvents::Union{Bool, Nothing}=nothing,
stateIndices=nothing,
valueIndices=nothing,
maxLabelLength=64,
plotkwargs...)

Plots data from a ME-FMU.

Optional t_in_solution controls if the first state in the solution is interpreted as t(ime).

Arguments

solution::FMUSolution: Struct contains information about the solution value, success, state and events of a specific FMU.

Keywords

states::Union{Bool, Nothing}: controls if states should be ploted (default = nothing)

values::Union{Bool, Nothing}: controls if values should be ploted (default = nothing)

timeEvents::Union{Bool, Nothing}=nothing: controls if timeEvents should be ploted (default = noting )

stateIndices=nothing: controls the number of ploted states

valueIndices=nothing: controls the number of ploted values

maxLabelLength=64: controls the maximum length for legend labels (too long labels are cut from front)

Returns

fig: Returns a figure containing the plotted data from a ME-FMU.

Save/load simulation results

FMI.fmiSaveSolution — Function

fmiSaveSolution(solution::FMUSolution, filepath::AbstractString)

saves the given solution in the file format specified by the ending of filepath. Currently .mat, .jld2 and .csv are supported.

Arguments

solution::FMUSolution: The simulation results that should be saved

filepath::AbstractString: The path specifing where to save the results, also indicating the file format. Supports *.mat, *.csv, *.JLD2

Keywords

keyword="solution": Specifies the type of data that should be saved

FMI.fmiSaveSolutionJLD2 — Function

fmiSaveSolutionJLD2(solution::FMUSolution, filepath::AbstractString; keyword="solution")

Saves a FMUSolution for later use. keyword is the used keyword for the jld data structure

FMI.fmiSaveSolutionMAT — Function

fmiSaveSolutionMAT(solution::FMUSolution, filepath::AbstractString)

Saves a FMUSolution to a mat file.

FMI.fmiSaveSolutionCSV — Function

fmiSaveSolutionCSV(solution::FMUSolution, filepath::AbstractString)

Saves a FMUSolution to a csv file.

FMI.fmiLoadSolution — Function

fmiLoadSolution(filepath::AbstractString; keyword="solution")

loads a previously saved solution from the file format specified by the ending of filepath. Currently .jld2 is supported.

Arguments

filepath::AbstractString: The path specifing where to save the results, also indicating the file format. Supports *.mat, *.csv, *.JLD2

keyword="solution": Specifies the type of data that should be saved

FMI.fmiLoadSolutionJLD2 — Function

fmiLoadSolutionJLD2(filepath::AbstractString; keyword="solution")

Loads a FMUSolution. Returns a previously saved FMUSolution. keyword is the used keyword for the jld data structure

FMI2 specific

FMI.fmi2Info — Function

 fmi2Info(fmu::FMU2)

Prints FMU related information.

Arguments

fmu::FMU2: Mutable struct representing a FMU and all it instantiated instances in the FMI 2.0.2 Standard.

Returns

Prints FMU related information.

Source

FMISpec2.0.2 Link: https://fmi-standard.org/

FMISpec2.0.2[p.22]: 2.1.4 Inquire Platform and Version Number of Header Files

FMISpec2.0.2[p.16]: 2.1.2 Platform Dependent Definitions

FMI.fmi2Simulate — Function

fmi2Simulate(fmu::FMU2, args...; kwargs...)

Wrapper for fmi2Simulate() in FMI/FMI2_sim.jl

fmi2Simulate(fmu::FMU2, c::Union{FMU2Component, Nothing}=nothing, tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing; kwargs...) Starts a simulation of the FMU instance for the matching FMU type, if both types are available, CS is preferred.

Keywords: - recordValues: Array of variables (strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues - setup: Boolean, if FMU should be setup (default=true) - reset: Boolean, if FMU should be reset before simulation (default reset=setup) - inputValues: Array of input variables (strings or variableIdentifiers) to set at every simulation step - inputFunction: Function to retrieve the values to set the inputs to - saveat: [ME only] Time points to save values at (interpolated) - solver: [ME only] Any Julia-supported ODE-solver (default is default from DifferentialEquations.jl) - customFx: [ME only, deprecated] Ability to give a custom state derivative function ẋ=f(x,t)

Returns: - success::Bool for CS-FMUs - ODESolution for ME-FMUs - if keyword recordValues is set, a tuple of type (success::Bool, DiffEqCallbacks.SavedValues) for CS-FMUs - if keyword recordValues is set, a tuple of type (ODESolution, DiffEqCallbacks.SavedValues) for ME-FMUs

FMI.fmi2SimulateME — Function

fmi2SimulateME(fmu::FMU2, args...; kwargs...)

Wrapper for fmi2SimulateME() in FMI/FMI2_sim.jl

function fmi2SimulateME(fmu::FMU2, c::Union{FMU2Component, Nothing}=nothing, tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing;
tolerance::Union{Real, Nothing} = nothing,
dt::Union{Real, Nothing} = nothing,
solver = nothing,
customFx = nothing,
recordValues::fmi2ValueReferenceFormat = nothing,
recordEventIndicators::Union{AbstractArray{<:Integer, 1}, UnitRange{<:Integer}, Nothing} = nothing,
recordEigenvalues::Bool=false,
saveat = nothing,
x0::Union{AbstractArray{<:Real}, Nothing} = nothing,
setup::Union{Bool, Nothing} = nothing,
reset::Union{Bool, Nothing} = nothing,
instantiate::Union{Bool, Nothing} = nothing,
freeInstance::Union{Bool, Nothing} = nothing,
terminate::Union{Bool, Nothing} = nothing,
inputValueReferences::fmi2ValueReferenceFormat = nothing,
inputFunction = nothing,
parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing,
dtmax::Union{Real, Nothing} = nothing,
callbacksBefore = [],
callbacksAfter = [],
showProgress::Bool = true,
kwargs...)

Simulates a FMU instance for the given simulation time interval. State- and Time-Events are handled correctly.

Via the optional keyword arguemnts inputValues and inputFunction, a custom input function f(c, u, t), f(c, t), f(u, t), f(c, u) or f(t) with c current component, u current state and t current time can be defined, that should return a array of values for fmi2SetReal(..., inputValues, inputFunction(...)).

Keywords:

solver: Any Julia-supported ODE-solver (default is the DifferentialEquations.jl default solver, currently AutoTsit5(Rosenbrock23()))

customFx: [deprecated] Ability to give a custom state derivative function ẋ=f(x,t)

recordValues: Array of variables (strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues

recordEventIndicators: Array or Range of event indicators (identified by integers) to record

recordEigenvalues: Boolean value, if eigenvalues shall be computed and recorded

saveat: Time points to save values at (interpolated)

setup: Boolean, if FMU should be setup (default=true)

reset: Union{Bool, :auto}, if FMU should be reset before simulation (default reset=:auto)

inputValueReferences: Array of input variables (strings or variableIdentifiers) to set at every simulation step

inputFunction: Function to retrieve the values to set the inputs to

parameters: Dictionary of parameter variables (strings or variableIdentifiers) and values (Real, Integer, Boolean, String) to set parameters during initialization

callbacks: custom callbacks to add

Returns:

If keyword recordValues is not set, a struct of type ODESolution.

If keyword recordValues is set, a tuple of type (ODESolution, DiffEqCallbacks.SavedValues).

FMI.fmi2SimulateCS — Function

fmi2SimulateCS(fmu::FMU2, args...; kwargs...)

Wrapper for fmi2SimulateCS() in FMI/FMI2_sim.jl

fmi2SimulateCS(fmu::FMU2, c::Union{FMU2Component, Nothing}=nothing, tspan::Union{Tuple{Float64, Float64}, Nothing}=nothing;
            tolerance::Union{Real, Nothing} = nothing,
            dt::Union{Real, Nothing} = nothing,
            recordValues::fmi2ValueReferenceFormat = nothing,
            saveat = [],
            setup::Union{Bool, Nothing} = nothing,
            reset::Union{Bool, Nothing} = nothing,
            instantiate::Union{Bool, Nothing} = nothing,
            freeInstance::Union{Bool, Nothing} = nothing,
            terminate::Union{Bool, Nothing} = nothing,
            inputValueReferences::fmi2ValueReferenceFormat = nothing,
            inputFunction = nothing,
            showProgress::Bool=true,
            parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing)

Starts a simulation of the Co-Simulation FMU instance.

Via the optional keyword arguments inputValues and inputFunction, a custom input function f(c, t) or f(t) with time t and component c can be defined, that should return a array of values for fmi2SetReal(..., inputValues, inputFunction(...)).

Keywords: - recordValues: Array of variables (strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues - saveat: Time points to save values at (interpolated) - setup: Boolean, if FMU should be setup (default=true) - reset: Boolean, if FMU should be reset before simulation (default reset=setup) - inputValueReferences: Array of input variables (strings or variableIdentifiers) to set at every simulation step - inputFunction: Function to retrieve the values to set the inputs to - parameters: Dictionary of parameter variables (strings or variableIdentifiers) and values (Real, Integer, Boolean, String) to set parameters during initialization Returns: - If keyword recordValues is not set, a boolean success is returned (simulation success). - If keyword recordValues is set, a tuple of type (true, DiffEqCallbacks.SavedValues) or (false, nothing).

FMI3 specific

FMI.fmi3Info — Function

Prints FMU related information.

FMI.fmi3Simulate — Function

fmi3Simulate(fmu::FMU3, args...; kwargs...)

Wrapper for fmi3Simulate() in FMI/FMI3_sim.jl

fmi3Simulate(fmu::FMU3, c::Union{FMU3Instance, Nothing}=nothing, t_start::Union{Real, Nothing} = nothing, t_stop::Union{Real, Nothing} = nothing; kwargs...)

Starts a simulation of the FMU instance for the matching FMU type, if multiple types are available, CS is preferred over ME, over SE.

Keywords: - recordValues: Array of variables (strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues - setup: Boolean, if FMU should be setup (default=true) - reset: Boolean, if FMU should be reset before simulation (default reset=setup) - inputValues: Array of input variables (strings or variableIdentifiers) to set at every simulation step - inputFunction: Function to retrieve the values to set the inputs to - saveat: Time points to save values at (interpolated) - solver: Any Julia-supported ODE-solver (default is default from DifferentialEquations.jl) - customFx: [ME only, deprecated] Ability to give a custom state derivative function ẋ=f(x,t)

Returns: - success::Bool for CS-FMUs - ODESolution for ME-FMUs - if keyword recordValues is set, a tuple of type (success::Bool, DiffEqCallbacks.SavedValues) for CS-FMUs - if keyword recordValues is set, a tuple of type (ODESolution, DiffEqCallbacks.SavedValues) for ME-FMUs

FMI.fmi3SimulateME — Function

fmi3SimulateME(fmu::FMU3, args...; kwargs...)

Wrapper for fmi3SimulateME() in FMI/FMI3_sim.jl

fmi3SimulateME(fmu::FMU3, c::Union{FMU3Instance, Nothing}=nothing, t_start::Union{Real, Nothing} = nothing, t_stop::Union{Real, Nothing} = nothing;
        tolerance::Union{Real, Nothing} = nothing,
        dt::Union{Real, Nothing} = nothing,
        solver = nothing,
        customFx = nothing,
        recordValues::fmi3ValueReferenceFormat = nothing,
        saveat = nothing,
        x0::Union{AbstractArray{<:Real}, Nothing} = nothing,
        setup::Union{Bool, Nothing} = nothing,
        reset::Union{Bool, Nothing} = nothing,
        instantiate::Union{Bool, Nothing} = nothing,
        freeInstance::Union{Bool, Nothing} = nothing,
        terminate::Union{Bool, Nothing} = nothing,
        inputValueReferences::fmi3ValueReferenceFormat = nothing,
        inputFunction = nothing,
        parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing,
        dtmax::Union{Real, Nothing} = nothing,
        callbacks = [],
        showProgress::Bool = true,
        kwargs...)

Simulates a FMU instance for the given simulation time interval. State- and Time-Events are handled correctly.

Via the optional keyword arguemnts inputValues and inputFunction, a custom input function f(c, u, t), f(c, t), f(u, t), f(c, u) or f(t) with c current instance, u current state and t current time can be defined, that should return a array of values for fmi3SetFloat64(..., inputValues, inputFunction(...)).

Keywords:

solver: Any Julia-supported ODE-solver (default is Tsit5)

customFx: [deperecated] Ability to give a custom state derivative function ẋ=f(x,t)

recordValues: Array of variables (strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues

saveat: Time points to save values at (interpolated)

setup: Boolean, if FMU should be setup (default=true)

reset: Union{Bool, :auto}, if FMU should be reset before simulation (default reset=:auto)

inputValueReferences: Array of input variables (strings or variableIdentifiers) to set at every simulation step

inputFunction: Function to retrieve the values to set the inputs to

parameters: Dictionary of parameter variables (strings or variableIdentifiers) and values (Real, Integer, Boolean, String) to set parameters during initialization

callbacks: custom callbacks to add

Returns:

If keyword recordValues is not set, a struct of type ODESolution.

If keyword recordValues is set, a tuple of type (ODESolution, DiffEqCallbacks.SavedValues).

FMI.fmi3SimulateSE — Function

fmi3SimulateSE(fmu::FMU3, c::Union{FMU3Instance, Nothing}=nothing, t_start::Union{Real, Nothing} = nothing, t_stop::Union{Real, Nothing} = nothing;
tolerance::Union{Real, Nothing} = nothing,
dt::Union{Real, Nothing} = nothing,
recordValues::fmi3ValueReferenceFormat = nothing,
saveat = [],
setup::Union{Bool, Nothing} = nothing,
reset::Union{Bool, Nothing} = nothing,
instantiate::Union{Bool, Nothing} = nothing,
freeInstance::Union{Bool, Nothing} = nothing,
terminate::Union{Bool, Nothing} = nothing,
inputValueReferences::fmi3ValueReferenceFormat = nothing,
inputFunction = nothing,
showProgress::Bool=true,
parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing)

TODO simulate ScheduledExecution not yet implemented in library

FMI.fmi3SimulateCS — Function

fmi3SimulateCS(fmu::FMU3, args...; kwargs...)

Wrapper for fmi3SimulateCS() in FMI/FMI3_sim.jl

fmi3SimulateCS(fmu::FMU3, c::Union{FMU3Instance, Nothing}=nothing, t_start::Union{Real, Nothing} = nothing, t_stop::Union{Real, Nothing} = nothing;
tolerance::Union{Real, Nothing} = nothing,
dt::Union{Real, Nothing} = nothing,
recordValues::fmi3ValueReferenceFormat = nothing,
saveat = [],
setup::Union{Bool, Nothing} = nothing,
reset::Union{Bool, Nothing} = nothing,
instantiate::Union{Bool, Nothing} = nothing,
freeInstance::Union{Bool, Nothing} = nothing,
terminate::Union{Bool, Nothing} = nothing,
inputValueReferences::fmi3ValueReferenceFormat = nothing,
inputFunction = nothing,
showProgress::Bool=true,
parameters::Union{Dict{<:Any, <:Any}, Nothing} = nothing)

Starts a simulation of the Co-Simulation FMU instance.

Via the optional keyword arguments inputValues and inputFunction, a custom input function f(c, t) or f(t) with time t and instance c can be defined, that should return a array of values for fmi3SetFloat64(..., inputValues, inputFunction(...)).

Keywords:

recordValues: Array of variables (strings or variableIdentifiers) to record. Results are returned as DiffEqCallbacks.SavedValues

saveat: Time points to save values at (interpolated)

setup: Boolean, if FMU should be setup (default=true)

reset: Boolean, if FMU should be reset before simulation (default reset=setup)

inputValueReferences: Array of input variables (strings or variableIdentifiers) to set at every simulation step

inputFunction: Function to retrieve the values to set the inputs to

parameters: Dictionary of parameter variables (strings or variableIdentifiers) and values (Real, Integer, Boolean, String) to set parameters during initialization

Returns:

If keyword recordValues is not set, a boolean success is returned (simulation success).

If keyword recordValues is set, a tuple of type (true, DiffEqCallbacks.SavedValues) or (false, nothing).